graduate vit from prototype
