+++ torchvision/models/__init__.py
@@ -10,6 +10,7 @@ from .mnasnet import *
 from .shufflenetv2 import *
 from .efficientnet import *
 from .regnet import *
+from .vision_transformer import *
 from . import detection
 from . import feature_extraction
 from . import optical_flow
